requirements.txt
avro==1.11.0
azure-common==1.1.28
azure-core==1.23.1
azure-eventhub==5.7.0
azure-functions==1.10.1
azure-identity==1.9.0
azure-schemaregistry==1.0.0
azure-storage-common==2.1.0
azure-storage-queue==2.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
databases==0.5.5
greenlet==1.1.2
idna==3.3
isodate==0.6.1
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
oauthlib==3.2.0
portalocker==2.4.0
pycparser==2.21f
PyJWT==2.3.0
python-dateutil==2.8.2
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.35
typing-extensions==4.2.0
uamqp==1.5.3
urllib3==1.26.9
psycopg2-binary==2.9.3
asyncpg